Implementing Notifications and Likes Functionality

+++ social_media_api/posts/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from rest_framework import viewsets, permissions, generics, status
 from .models import Post, Comment
 from .serializers import PostSerializer, CommentSerializer
@@ -8,6 +8,7 @@ from rest_framework.permissions import IsAuthenticated
 from .models import Post, Like
 from notifications.models import Notification
 from rest_framework.response import Response
+from django.contrib.contenttypes.models import ContentType
 
 User = get_user_model()
 # Create your views here.
@@ -42,30 +43,35 @@ class FeedView(generics.ListAPIView):
 class LikePostView(generics.GenericAPIView):
     permission_classes = [permissions.IsAuthenticated]
 
-    def post(self, request, post_id):
-        post = Post.objects.get(id=post_id)
-        like, created = Like.objects.get_or_create(user=request.user, post=post)
+    def post(self, request, pk):
+        post = get_object_or_404(Post, pk=pk)
+        user = request.user
+
+        like, created = Like.objects.get_or_create(user=user, post=post)
         if not created:
-            return Response({'detail': 'You have liked this post. '}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': 'You have liked this post. '}, status=status.HTTP_400_BAD_REQUEST)
         
         if request.user != post.author:
             Notification.objects.create(
                 recipient=post.author,
-                actor=request.user,
+                actor=user,
                 verb='liked',
-                target=post
+                target_content_type=ContentType.objects.get_for_model(post),
+                target_object_id=post.id
             )
-        return Response({'detail': 'Post liked.'}, status=status.HTTP_201_CREATED)
+        return Response({'message': 'Post liked.'}, status=status.HTTP_201_CREATED)
     
 class UnlikePostView(generics.GenericAPIView):
     permission_classes = [permissions.IsAuthenticated]
 
-    def post(self, request, post_id):
-        post = Post.objects.get(id=post_id)
-        try:
-            like = Like.objects.get(user=request.user, post=post)
-            like.delete()
-            return Response({'detail': 'Post unliked.'}, status=status.HTTP_200_OK)
-        except Like.DoesNotExist:
-            return Response({'detail': 'You have not liked this post.'}, status=status.HTTP_400_BAD_REQUEST)
+    def delete(self, request, pk):
+        post = get_object_or_404(Post, pk=pk)
+        user = request.user
+        
+        like = Like.objects.get(user=user, post=post).first()
+        if not like:
+            return Response({'message': 'You have not liked this post'}, status=status.HTTP_400_BAD_REQUEST)
+        like.delete()
+        
+        return Response({'message': 'Post unliked..'}, status=status.HTTP_204_NO_CONTENT)
